+++ fast-slic-avx2.h
@@ -8,7 +8,7 @@
 extern "C" {
 #endif
     void fast_slic_initialize_clusters_avx2(int H, int W, int K, const uint8_t* image, Cluster *clusters);
-    void fast_slic_iterate_avx2(int H, int W, int K, uint32_t compactness, uint8_t quantize_level, int max_iter, const uint8_t* image, Cluster* clusters, uint32_t* assignment);
+    void fast_slic_iterate_avx2(int H, int W, int K, float compactness, uint8_t quantize_level, int max_iter, const uint8_t* image, Cluster* clusters, uint32_t* assignment);
     int fast_slic_supports_avx2();
 #ifdef __cplusplus
 }
